# api_spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LMS API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/lessons/": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get Lessons",
                "description": "Get all lessons",
                "operationId": "get_lessons_api_lessons__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Lesson"
                                    },
                                    "type": "array",
                                    "title": "Response Get Lessons Api Lessons  Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Create Lesson",
                "description": "Create a new lesson",
                "operationId": "create_lesson_api_lessons__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LessonRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lessons/{lesson_id}": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get Lesson",
                "description": "Get a specific lesson by ID",
                "operationId": "get_lesson_api_lessons__lesson_id__get",
                "parameters": [
                    {
                        "name": "lesson_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Lesson Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/exercises/": {
            "get": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Get Exercises",
                "description": "Get all exercises.\nIf lesson_id is provided, filters exercises by that lesson_id.",
                "operationId": "get_exercises_api_exercises__get",
                "parameters": [
                    {
                        "name": "lesson_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Lesson Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Exercise"
                                    },
                                    "title": "Response Get Exercises Api Exercises  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Create Exercise",
                "description": "Create a new exercise",
                "operationId": "create_exercise_api_exercises__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExerciseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exercise"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/exercises/{exercise_id}": {
            "get": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Get Exercise",
                "description": "Get a specific exercise by ID",
                "operationId": "get_exercise_api_exercises__exercise_id__get",
                "parameters": [
                    {
                        "name": "exercise_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Exercise Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Exercise"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/exercises/batch/": {
            "post": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Create Exercises Batch",
                "description": "Create multiple exercises at once using a template",
                "operationId": "create_exercises_batch_api_exercises_batch__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExerciseTemplate-Input"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Exercise"
                                    },
                                    "type": "array",
                                    "title": "Response Create Exercises Batch Api Exercises Batch  Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Exercise": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "exercise_type": {
                        "type": "string",
                        "title": "Exercise Type"
                    },
                    "content": {
                        "type": "object",
                        "title": "Content"
                    },
                    "correct_answer": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Correct Answer"
                    },
                    "max_score": {
                        "type": "integer",
                        "title": "Max Score",
                        "default": 1
                    },
                    "grading_type": {
                        "type": "string",
                        "title": "Grading Type",
                        "default": "auto"
                    },
                    "lesson_id": {
                        "type": "integer",
                        "title": "Lesson Id"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "correct_order": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Correct Order"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "exercise_type",
                    "content",
                    "correct_answer",
                    "lesson_id",
                    "id"
                ],
                "title": "Exercise"
            },
...
            "ExerciseRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "content": {
                        "type": "object",
                        "title": "Content"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "description",
                    "content"
                ],
                "title": "ExerciseRequest"
            },
...
            "ExerciseTemplate-Input": {
                "properties": {
                    "lesson_id": {
                        "type": "integer",
                        "title": "Lesson Id"
                    },
                    "exercises": {
                        "items": {
                            "$ref": "#/components/schemas/ExerciseRequest"
                        },
                        "type": "array",
                        "title": "Exercises"
                    }
                },
                "type": "object",
                "required": [
                    "lesson_id",
                    "exercises"
                ],
                "title": "ExerciseTemplate"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Lesson": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "description",
                    "content",
                    "id"
                ],
                "title": "Lesson"
            },
            "LessonRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "order": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Order"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "description",
                    "content"
                ],
                "title": "LessonRequest"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}